@@ -2,82 +2,55 @@
 
 namespace App\Controller\Api;
 
+use App\Controller\AbstractRestController;
 use App\Entity\User;
 use App\Service\Response\MessageResponseService;
 use App\Service\SocketService;
-use FOS\RestBundle\Controller\AbstractFOSRestController;
-use FOS\RestBundle\Controller\Annotations as Rest;
-use FOS\RestBundle\View\View;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
-/**
- * @Route("/messages")
- */
-class MessagesController extends AbstractFOSRestController
+#[Route(path: '/messages')]
+class MessagesController extends AbstractRestController
 {
-    /**
-     * @Route("/{offset}", name="messages_index")
-     * @Route("/group-list/{offset}", name="messages")
-     * @IsGranted("ROLE_USER")
-     * @Rest\View(serializerGroups={"message","user"})
-     * @param MessageResponseService $messageResponseService
-     * @param int $offset
-     * @return View
-     */
+    #[Route(path: '/{offset}', name: 'messages_index')]
+    #[Route(path: '/group-list/{offset}', name: 'messages')]
+    #[IsGranted('ROLE_USER')]
     public function groupList(
         MessageResponseService $messageResponseService,
         int $offset = 0
-    ): View
-    {
-        /**@var User $sessionUser*/
+    ): JsonResponse {
+        /** @var User $sessionUser */
         $sessionUser = $this->getUser();
         $response = $messageResponseService->groupList($sessionUser, $offset);
 
-        return View::create($response);
+        return $this->json($response, context: ['groups' => ["message", "user"]]);
     }
 
-    /**
-     * @Route("/detail/{user}/{offset}", name="messages_detail")
-     * @IsGranted("ROLE_USER")
-     * @Rest\View(serializerGroups={"message","user"})
-     * @param MessageResponseService $messageResponseService
-     * @param User $user
-     * @param int $offset
-     * @return View
-     */
+    #[Route(path: '/detail/{user}/{offset}', name: 'messages_detail')]
+    #[IsGranted('ROLE_USER')]
     public function detail(
         MessageResponseService $messageResponseService,
         User $user,
         int $offset = 0
-    ): View
-    {
-        /**@var User $sessionUser*/
+    ): JsonResponse {
+        /** @var User $sessionUser */
         $sessionUser = $this->getUser();
         $response = $messageResponseService->detail($sessionUser, $user, $offset);
 
-        return View::create($response);
+        return $this->json($response, context: ['groups' => ["message", "user"]]);
     }
 
-    /**
-     * @Route("/add/{user}", name="messages_add", methods={"POST"})
-     * @IsGranted("ROLE_USER")
-     * @Rest\View(serializerGroups={"message","user"})
-     * @param User $user
-     * @param Request $request
-     * @param SocketService $socketService
-     * @param MessageResponseService $messageResponseService
-     * @return View
-     */
+    #[Route(path: '/add/{user}', name: 'messages_add', methods: ['POST'])]
+    #[IsGranted('ROLE_USER')]
     public function add(
         User $user,
         Request $request,
         SocketService $socketService,
         MessageResponseService $messageResponseService
-    ): View
-    {
-        /**@var User $sessionUser*/
+    ): JsonResponse {
+        /** @var User $sessionUser */
         $sessionUser = $this->getUser();
         $text = $request->get('text');
 
@@ -85,31 +58,22 @@ class MessagesController extends AbstractFOSRestController
 
         $socketService->sendMessage($user, $response);
 
-        return  View::create($response);
+        return $this->json($response, context: ['groups' => ["message", "user"]]);
     }
 
-    /**
-     * @Route("/read/{user}", name="messages_read", methods={"GET"})
-     * @IsGranted("ROLE_USER")
-     * @Rest\View(serializerGroups={"message","user"})
-     * @param User $user
-     * @param SocketService $socketService
-     * @param MessageResponseService $messageResponseService
-     * @return View
-     */
+    #[Route(path: '/read/{user}', name: 'messages_read', methods: ['GET'])]
+    #[IsGranted('ROLE_USER')]
     public function read(
         User $user,
         SocketService $socketService,
         MessageResponseService $messageResponseService
-    ): View
-    {
-        /**@var User $sessionUser*/
+    ): JsonResponse {
+        /** @var User $sessionUser */
         $sessionUser = $this->getUser();
         $result = $messageResponseService->read($sessionUser, $user);
 
         $socketService->sendRead($user, $sessionUser);
 
-        return View::create($result);
+        return $this->json($result, context: ['groups' => ["message", "user"]]);
     }
-
 }
