@@ -1,163 +1,101 @@
 <?php
 
-
 namespace App\Document;
 
 use Doctrine\ODM\MongoDB\Mapping\Annotations as MongoDB;
 use Symfony\Component\Serializer\Annotation\Groups;
+use DateTimeInterface;
 
-/**
- * @MongoDB\Document(collection="message")
- */
+#[MongoDB\Document(collection: "message")]
 class Message
 {
-    /**
-     * @MongoDB\Id
-     * @Groups({"message"})
-     */
-    protected $id;
+    #[MongoDB\Id]
+    #[Groups(["message"])]
+    private ?string $id;
 
-    /**
-     * @var int
-     * @MongoDB\Field(type="integer")
-     * @Groups({"message"})
-     */
-    protected $from;
+    #[MongoDB\Field(type: "int")]
+    #[Groups(["message"])]
+    private ?int $from;
 
-    /**
-     * @var int
-     * @MongoDB\Field(type="integer")
-     * @Groups({"message"})
-     */
-    protected $to;
+    #[MongoDB\Field(type: "int")]
+    #[Groups(["message"])]
+    private ?int $to;
 
-    /**
-     * @var string
-     * @MongoDB\Field(type="string")
-     * @Groups({"message"})
-     */
-    private $text;
+    #[MongoDB\Field(type: "string")]
+    #[Groups(["message"])]
+    private ?string $text;
 
-    /**
-     * @var \DateTimeInterface
-     * @MongoDB\Field(type="date")
-     * @Groups({"message"})
-     */
-    private $date;
+    #[MongoDB\Field(type: "date")]
+    #[Groups(["message"])]
+    private ?DateTimeInterface $date;
 
-    /**
-     * @var bool
-     * @MongoDB\Field (type="boolean")
-     * @Groups({"message"})
-     */
-    private $read;
+    #[MongoDB\Field(type: "bool")]
+    #[Groups(["message"])]
+    private ?bool $read;
 
-    /**
-     * @return mixed
-     */
-    public function getId()
+    public function getId(): ?string
     {
         return $this->id;
     }
 
-    /**
-     * @param mixed $id
-     * @return Message
-     */
-    public function setId($id)
+    public function setId(string $id): self
     {
         $this->id = $id;
         return $this;
     }
 
-    /**
-     * @return int
-     */
     public function getFrom(): int
     {
         return $this->from;
     }
 
-    /**
-     * @param int $from
-     * @return Message
-     */
-    public function setFrom(int $from): Message
+    public function setFrom(int $from): self
     {
         $this->from = $from;
         return $this;
     }
 
-    /**
-     * @return int
-     */
     public function getTo(): int
     {
         return $this->to;
     }
 
-    /**
-     * @param int $to
-     * @return Message
-     */
-    public function setTo(int $to): Message
+    public function setTo(int $to): self
     {
         $this->to = $to;
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getText(): string
     {
         return $this->text;
     }
 
-    /**
-     * @param string $text
-     * @return Message
-     */
-    public function setText(string $text): Message
+    public function setText(string $text): self
     {
         $this->text = $text;
         return $this;
     }
 
-    /**
-     * @return \DateTimeInterface
-     */
-    public function getDate(): \DateTimeInterface
+    public function getDate(): DateTimeInterface
     {
         return $this->date;
     }
 
-    /**
-     * @param \DateTimeInterface $date
-     * @return Message
-     */
-    public function setDate(\DateTimeInterface $date): Message
+    public function setDate(DateTimeInterface $date): self
     {
         $this->date = $date;
         return $this;
     }
 
-    /**
-     * @return bool
-     */
     public function isRead(): ?bool
     {
         return $this->read;
     }
 
-    /**
-     * @param bool $read
-     * @return Message
-     */
-    public function setRead(bool $read): Message
+    public function setRead(?bool $read): self
     {
         $this->read = $read;
         return $this;
     }
-
 }
