@@ -1,8 +1,6 @@
 <?php
 
-
 namespace App\Type\Story;
-
 
 use App\Document\Result\StoryGroup;
 use App\Entity\User;
@@ -10,83 +8,50 @@ use Symfony\Component\Serializer\Annotation\Groups;
 
 class StoryGroupResponse
 {
-    /**
-     * @var User|null
-     * @Groups({"story"})
-     */
-    private $user;
+    #[Groups(["story"])]
+    private ?User $user = null; // Property promotion ve nullability
 
-    /**
-     * @var StoryGroupItemResponse[]
-     * @Groups({"story"})
-     */
-    private $items;
+    #[Groups(["story"])]
+    private array $items = []; // Property promotion
 
-    /**
-     * @return User|null
-     */
     public function getUser(): ?User
     {
         return $this->user;
     }
 
-    /**
-     * @param User|null $user
-     * @return self
-
-     */
     public function setUser(?User $user): self
     {
         $this->user = $user;
         return $this;
     }
 
-    /**
-     * @return self[]
-     */
     public function getItems(): array
     {
         return $this->items;
     }
 
-    /**
-     * @param StoryGroupResponse[] $items
-     * @return self
-     */
     public function setItems(array $items): self
     {
         $this->items = $items;
         return $this;
     }
 
-    /**
-     * @param StoryGroupItemResponse $item
-     * @return self
-     */
     public function addItems(StoryGroupItemResponse $item): self
     {
         $this->items[] = $item;
         return $this;
     }
 
-    /**
-     * @param StoryGroup $groupItem
-     * @param User $user
-     * @return self
-     */
-    public static function fill(StoryGroup $groupItem, User $user)
+    public static function fill(StoryGroup $groupItem, User $user): self // Return type ekle
     {
         $self = new self();
         $self->setUser($user);
 
         foreach ($groupItem->getItems() as $item) {
-
             $tItem = StoryGroupItemResponse::fill($item, $user);
-
             $self->addItems($tItem);
         }
 
         return $self;
     }
-
 }
