@@ -1,10 +1,7 @@
 <?php
 
-
 namespace App\Type\Story;
 
-
-use App\Document\Result\StoryGroup;
 use App\Document\Result\StoryMeItem;
 use App\Document\Result\StoryViewItem;
 use App\Entity\User;
@@ -12,205 +9,125 @@ use Symfony\Component\Serializer\Annotation\Groups;
 
 class StoryMeItemResponse
 {
-    /**
-     * @var string
-     * @Groups({"story"})
-     */
-    protected $id;
+    #[Groups(["story"])]
+    protected string $id;
 
-    /**
-     * @var User
-     * @Groups({"story"})
-     */
-    private $user;
+    #[Groups(["story"])]
+    private User $user;
 
-    /**
-     * @var string
-     * @Groups({"story"})
-     */
-    private $rootPath;
+    #[Groups(["story"])]
+    private string $rootPath;
 
-    /**
-     * @var string
-     * @Groups({"story"})
-     */
-    private $path;
+    #[Groups(["story"])]
+    private string $path;
 
-    /**
-     * @var string
-     * @Groups({"story"})
-     */
-    private $fileName;
+    #[Groups(["story"])]
+    private string $fileName;
 
-    /**
-     * @var \DateTimeInterface
-     * @Groups({"story"})
-     */
-    private $date;
+    #[Groups(["story"])]
+    private \DateTimeInterface $date;
 
-    /**
-     * @var StoryViewItem[]
-     * @Groups({"story"})
-     */
-    private $views = [];
+    #[Groups(["story"])]
+    private array $views = []; // Başlangıç değeri eklendi
 
-    /**
-     * @var int
-     * @Groups({"story"})
-     */
-    private $viewsLength;
+    #[Groups(["story"])]
+    private int $viewsLength;
 
-    /**
-     * @return string
-     */
-    public function getId()
+    public function getId(): string
     {
         return $this->id;
     }
 
-    /**
-     * @param string $id
-     * @return self
-     */
-    public function setId( $id): self
+    public function setId(string $id): self
     {
         $this->id = $id;
         return $this;
     }
 
-    /**
-     * @return User
-     */
     public function getUser(): User
     {
         return $this->user;
     }
 
-    /**
-     * @param User $user
-     * @return self
-     */
     public function setUser(User $user): self
     {
         $this->user = $user;
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getRootPath(): string
     {
         return $this->rootPath;
     }
 
-    /**
-     * @param string $rootPath
-     * @return self
-     */
     public function setRootPath(string $rootPath): self
     {
         $this->rootPath = $rootPath;
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getPath(): string
     {
         return $this->path;
     }
 
-    /**
-     * @param string $path
-     * @return self
-     */
     public function setPath(string $path): self
     {
         $this->path = $path;
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getFileName(): string
     {
         return $this->fileName;
     }
 
-    /**
-     * @param string $fileName
-     * @return self
-     */
     public function setFileName(string $fileName): self
     {
         $this->fileName = $fileName;
         return $this;
     }
 
-    /**
-     * @return \DateTimeInterface
-     */
     public function getDate(): \DateTimeInterface
     {
         return $this->date;
     }
 
-    /**
-     * @param \DateTimeInterface $date
-     * @return self
-     */
     public function setDate(\DateTimeInterface $date): self
     {
         $this->date = $date;
         return $this;
     }
 
-    /**
-     * @return StoryViewItemResponse[]
-     */
     public function getViews(): array
     {
         return $this->views;
     }
 
-    /**
-     * @param StoryViewItemResponse[] $views
-     * @return self
-     */
     public function setViews(array $views): self
     {
         $this->views = $views;
         return $this;
     }
 
-    public function addView(StoryViewItemResponse $viewItemResponse)
+    public function addView(StoryViewItemResponse $viewItemResponse): self // Return type eklendi
     {
         $this->views[] = $viewItemResponse;
         return $this;
     }
 
-    /**
-     * @return int
-     */
     public function getViewsLength(): int
     {
         return $this->viewsLength;
     }
 
-    /**
-     * @param int $viewsLength
-     * @return self
-     */
     public function setViewsLength(int $viewsLength): self
     {
         $this->viewsLength = $viewsLength;
         return $this;
     }
 
-    public static function fill(User $user, StoryMeItem $storyMeItem)
+    public static function fill(User $user, StoryMeItem $storyMeItem): self // Return type eklendi
     {
         $self = new self();
         $self
