@@ -3,39 +3,35 @@
 namespace App\Service;
 
 use App\Entity\User;
-use App\Repository\UserRepository;
 use App\Type\Image\Uploaded;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 
 class UserService
 {
     private EntityManagerInterface $entityManager;
-    private UserRepository $userRepository;
     private ImageUploadService $imageUploadService;
     private ParameterBagInterface $parameterBag;
-    private UserPasswordEncoderInterface $userPasswordEncoder;
+    private UserPasswordHasherInterface $passwordHasher;
 
     public function __construct(
         EntityManagerInterface $entityManager,
-        UserRepository $userRepository,
         ImageUploadService $imageUploadService,
         ParameterBagInterface $parameterBag,
-        UserPasswordEncoderInterface $userPasswordEncoder
+        UserPasswordHasherInterface $passwordHasher
     )
     {
         $this->entityManager = $entityManager;
-        $this->userRepository = $userRepository;
         $this->imageUploadService = $imageUploadService;
         $this->parameterBag = $parameterBag;
-        $this->userPasswordEncoder = $userPasswordEncoder;
+        $this->passwordHasher = $passwordHasher;
     }
 
     public function create(User $user)
     {
         $user->setRoles(['ROLE_USER']);
-        $user->setPassword($this->userPasswordEncoder->encodePassword($user, $user->getPassword()));
+        $user->setPassword($this->passwordHasher->hashPassword($user, $user->getPassword()));
         $user->setLastSeen(new \DateTime());
         $this->entityManager->persist($user);
         $this->entityManager->flush();
